+++ tasks/seq/kholin_k_multidimensional_integrals_rectangle/src/ops_seq.cpp
@@ -1,11 +1,13 @@
 #include "seq/kholin_k_multidimensional_integrals_rectangle/include/ops_seq.hpp"
 
+#include <cmath>
+#include <cstddef>
 #include <functional>
 #include <vector>
 
 double kholin_k_multidimensional_integrals_rectangle_seq::TestTaskSequential::Integrate(
     const Function& f, const std::vector<double>& l_limits, const std::vector<double>& u_limits,
-    const std::vector<double>& h, std::vector<double>& f_values, size_t curr_index_dim, size_t dim, size_t n) {
+    const std::vector<double>& h, std::vector<double>& f_values, double curr_index_dim, double dim, double n) {
   if (curr_index_dim == dim) {
     return f(f_values);
   }
@@ -20,10 +22,10 @@ double kholin_k_multidimensional_integrals_rectangle_seq::TestTaskSequential::In
 
 double kholin_k_multidimensional_integrals_rectangle_seq::TestTaskSequential::IntegrateWithRectangleMethod(
     const Function& f, std::vector<double>& f_values, const std::vector<double>& l_limits,
-    const std::vector<double>& u_limits, size_t dim, size_t n) {
+    const std::vector<double>& u_limits, double dim, double n) {
   std::vector<double> h(dim);
   for (size_t i = 0; i < dim; ++i) {
-    h[i] = (u_limits[i] - l_limits[i]) / static_cast<double>(n);
+    h[i] = (u_limits[i] - l_limits[i]) / n;
   }
 
   return Integrate(f, l_limits, u_limits, h, f_values, 0, dim, n);
@@ -31,7 +33,7 @@ double kholin_k_multidimensional_integrals_rectangle_seq::TestTaskSequential::In
 
 double kholin_k_multidimensional_integrals_rectangle_seq::TestTaskSequential::RunMultistepSchemeMethodRectangle(
     const Function& f, std::vector<double>& f_values, const std::vector<double>& l_limits,
-    const std::vector<double>& u_limits, size_t dim, int n, double epsilon) {
+    const std::vector<double>& u_limits, double dim, double n, double epsilon) {
   double i_n = IntegrateWithRectangleMethod(f, f_values, l_limits, u_limits, dim, n);
   double i_2n = 0.0;
   double delta = 0;
@@ -52,7 +54,7 @@ bool kholin_k_multidimensional_integrals_rectangle_seq::TestTaskSequential::PreP
   sz_lower_limits_ = task_data->inputs_count[1];
   sz_upper_limits_ = task_data->inputs_count[2];
 
-  auto* ptr_dim = reinterpret_cast<size_t*>(task_data->inputs[0]);
+  auto* ptr_dim = reinterpret_cast<double*>(task_data->inputs[0]);
   dim_ = *ptr_dim;
 
   auto* ptr_f_values = reinterpret_cast<double*>(task_data->inputs[1]);
@@ -70,7 +72,7 @@ bool kholin_k_multidimensional_integrals_rectangle_seq::TestTaskSequential::PreP
   auto* ptr_epsilon = reinterpret_cast<double*>(task_data->inputs[5]);
   epsilon_ = *ptr_epsilon;
 
-  auto* ptr_start_n = reinterpret_cast<int*>(task_data->inputs[6]);
+  auto* ptr_start_n = reinterpret_cast<double*>(task_data->inputs[6]);
   start_n_ = *ptr_start_n;
 
   result_ = 0.0;
